[Northamptonshire] ignore linked defects until scheduled
The flow from inspection to defect goes:
1: inspection created
2: inspection assigned -> FMS to investigating
3: inspection closed, set to works instructed -> FMS action scheduled
4: defect created
5: defect pulled in to Keir, defect set to order raised
6: Fixed, defect to remedied -> FMS fixed

At step 4 the defect is confirmed which maps to investigating. We want
to ignore this step for defects linked to inspections as othwise the FMS
status goes

investigating
action scheduled
investigating
action scheduled

which is confusing.

However, we want non linked defects to follow a normal course.

+++ perllib/Open311/Endpoint/Integration/Alloy.pm
@@ -277,6 +277,7 @@ sub get_service_request_updates {
         my $status = 'open';
         my $description = '';
         my $fms_id = '';
+        my $linked_defect;
         my @attributes = @{$update->{values}};
         for my $att (@attributes) {
             # these might be specific to each design so will probably need
@@ -293,6 +294,7 @@ sub get_service_request_updates {
             }
 
             if ($att->{attributeCode} =~ /_FIXMYSTREET_ID$/) {
+                $linked_defect = 1;
                 $fms_id = $att->{value};
             }
         }
@@ -305,9 +307,13 @@ sub get_service_request_updates {
         for my $parent (@$parents) {
             next unless $parent->{actualParentSourceTypeId} == $self->config->{defect_inspection_parent_id}; # request for service
 
+            $linked_defect = 1;
             $service_request_id = $parent->{parentResId};
             $fms_id = undef;
         }
+
+        # we don't care about linked defects until they have been scheduled
+        next if $linked_defect && ( $status eq 'open' || $status eq 'investigating' );
 
         my $update_time = $self->get_time_for_version($update->{resourceId}, $update->{version}->{resourceSystemVersionId});
         my $update_dt = DateTime::Format::W3CDTF->new->parse_datetime( $update_time )->truncate( to => 'second' );
